# Memcached: preallocate slice used for grouping cache keys by server