**What this PR does**:

We know the number of keys and so should pre-allocate the slice used. Runtime of both the new and old version varies by a lot however, the new version tends to be faster (in addition to consistently allocating less memory).

```
$ benchstat -delta-test=none old.txt new.txt 
name                                 old time/op    new time/op    delta
MemcachedClient_sortKeysByServer-16    2.53ms ± 0%    2.11ms ± 0%  -16.60%

name                                 old alloc/op   new alloc/op   delta
MemcachedClient_sortKeysByServer-16    1.31MB ± 0%    0.89MB ± 0%  -32.23%

name                                 old allocs/op  new allocs/op  delta
MemcachedClient_sortKeysByServer-16     20.1k ± 0%     20.1k ± 0%   -0.02%
```

**Which issue(s) this PR fixes**:

N/A

**Checklist**
- [X] Tests updated
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`